Hibernate config cleanup
