+++ javers-hibernate/src/test/java/org/javers/hibernate/integration/config/HibernateConfig.java
@@ -1,0 +1,7 @@
+package org.javers.hibernate.integration.config;
+
+/**
+ * Created by bartosz.galek on 15.06.15.
+ */
+public class HibernateConfig {
+}
