+++ javers-hibernate/src/test/java/org/javers/hibernate/integration/config/JaversFieldProxyManagerConfig.java
@@ -1,0 +1,46 @@
+package org.javers.hibernate.integration.config;
+
+import org.javers.core.Javers;
+import org.javers.hibernate.HibernateProxyManager;
+import org.javers.repository.sql.DialectName;
+import org.javers.repository.sql.JaversSqlRepository;
+import org.javers.repository.sql.SqlRepositoryBuilder;
+import org.javers.spring.jpa.JpaHibernateConnectionProvider;
+import org.javers.spring.jpa.TransactionalJaversBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+/**
+ * @author bartosz walacik
+ */
+@Configuration
+@ComponentScan(basePackages = "org.javers.hibernate")
+@EnableTransactionManagement
+@EnableAspectJAutoProxy
+@EnableJpaRepositories(basePackages = "org.javers.hibernate")
+public class JaversFieldProxyManagerConfig {
+
+    //.. JaVers setup ..
+
+    /**
+     * Creates JaVers instance with {@link JaversSqlRepository}
+     */
+    @Bean
+    public Javers javers(JpaHibernateConnectionProvider jpaHibernateConnectionProvider) {
+        JaversSqlRepository sqlRepository = SqlRepositoryBuilder
+                .sqlRepository()
+                .withConnectionProvider(jpaHibernateConnectionProvider)
+                .withDialect(DialectName.H2)
+                .build();
+
+        return TransactionalJaversBuilder
+                .javers()
+                .registerJaversRepository(sqlRepository)
+                .withGraphFactoryHook(new HibernateProxyManager())
+                .build();
+    }
+}
