--- javers-hibernate/src/test/java/org/javers/hibernate/integration/config/JaversProxyManagerApplicationConfig.java
@@ -1,149 +1,0 @@
-package org.javers.hibernate.integration.config;
-
-import org.javers.core.Javers;
-import org.javers.hibernate.HibernateProxyManager;
-import org.javers.repository.sql.ConnectionProvider;
-import org.javers.repository.sql.DialectName;
-import org.javers.repository.sql.JaversSqlRepository;
-import org.javers.repository.sql.SqlRepositoryBuilder;
-import org.javers.spring.annotation.JaversAuditable;
-import org.javers.spring.annotation.JaversSpringDataAuditable;
-import org.javers.spring.auditable.AuthorProvider;
-import org.javers.spring.auditable.aspect.JaversAuditableRepositoryAspect;
-import org.javers.spring.jpa.JpaHibernateConnectionProvider;
-import org.javers.spring.jpa.TransactionalJaversBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.JpaVendorAdapter;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-import javax.persistence.EntityManagerFactory;
-import javax.sql.DataSource;
-import java.util.Properties;
-
-/**
- * @author bartosz walacik
- */
-@Configuration
-@ComponentScan(basePackages = "org.javers.hibernate")
-@EnableTransactionManagement
-@EnableAspectJAutoProxy
-@EnableJpaRepositories(basePackages = "org.javers.hibernate")
-public class JaversProxyManagerApplicationConfig {
-
-    //.. JaVers setup ..
-
-    /**
-     * Creates JaVers instance with {@link JaversSqlRepository}
-     */
-    @Bean
-    public Javers javers() {
-        JaversSqlRepository sqlRepository = SqlRepositoryBuilder
-                .sqlRepository()
-                .withConnectionProvider(jpaConnectionProvider())
-                .withDialect(DialectName.H2)
-                .build();
-
-        return TransactionalJaversBuilder
-                .javers()
-                .registerJaversRepository(sqlRepository)
-                .withGraphFactoryHook(new HibernateProxyManager())
-                .build();
-    }
-
-    /**
-     * Integrates {@link JaversSqlRepository} with Spring {@link JpaTransactionManager}
-     */
-    @Bean
-    public ConnectionProvider jpaConnectionProvider() {
-        return new JpaHibernateConnectionProvider();
-    }
-    //.. EOF JaVers setup ..
-
-
-    //.. Spring-JPA-Hibernate setup ..
-    @Bean
-    public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
-        LocalContainerEntityManagerFactoryBean em = new LocalContainerEntityManagerFactoryBean();
-        em.setDataSource(dataSource());
-        em.setPackagesToScan("org.javers.hibernate");
-
-        JpaVendorAdapter vendorAdapter = new HibernateJpaVendorAdapter();
-        em.setJpaVendorAdapter(vendorAdapter);
-        em.setJpaProperties(additionalProperties());
-
-        return em;
-    }
-
-    @Bean
-    public PlatformTransactionManager transactionManager(EntityManagerFactory emf) {
-        JpaTransactionManager transactionManager = new JpaTransactionManager();
-        transactionManager.setEntityManagerFactory(emf);
-
-        return transactionManager;
-    }
-
-    @Bean
-    public PersistenceExceptionTranslationPostProcessor exceptionTranslation() {
-        return new PersistenceExceptionTranslationPostProcessor();
-    }
-
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName("org.h2.Driver");
-        dataSource.setUrl("jdbc:h2:mem:test;DB_CLOSE_DELAY=-1");
-        return dataSource;
-    }
-
-    /**
-     * Enables Repository auto-audit aspect. <br/>
-     *
-     * Use {@link JaversSpringDataAuditable}
-     * to annotate Spring Data Repositories
-     * or {@link JaversAuditable} for ordinary Repositories.
-     */
-    @Bean
-    public JaversAuditableRepositoryAspect javersAuditableRepositoryAspect() {
-        return new JaversAuditableRepositoryAspect(javers(), authorProvider());
-    }
-
-    /**
-     * Required by Repository auto-audit aspect. <br/><br/>
-     *
-     * Returns mock implementation for testing.
-     * <br/>
-     * Provide real implementation,
-     * when using Spring Security you can use
-     * {@link org.javers.spring.auditable.SpringSecurityAuthorProvider}.
-     */
-    @Bean
-    public AuthorProvider authorProvider() {
-        return new AuthorProvider() {
-            @Override
-            public String provide() {
-                return "unknown";
-            }
-        };
-    }
-
-    Properties additionalProperties() {
-        Properties properties = new Properties();
-        properties.setProperty("hibernate.hbm2ddl.auto", "create");
-        properties.setProperty("hibernate.connection.autocommit", "false");
-        properties.setProperty("hibernate.dialect", "org.hibernate.dialect.H2Dialect");
-        properties.setProperty("hibernate.current_session_context_class", "thread");
-        properties.setProperty("hibernate.enable_lazy_load_no_trans", "true");
-        return properties;
-    }
-    //.. EOF Spring-JPA-Hibernate setup ..
-}
